src/requirements.txt
flask
google-cloud-compute
discord-interactions
functions-framework
requests